+++ stdlib.go
@@ -67,7 +67,7 @@ var typeOfFunc = function.New(&function.Spec{
 	},
 	Type: function.StaticReturnType(cty.String),
 	Impl: func(args []cty.Value, retType cty.Type) (cty.Value, error) {
-		return cty.StringVal(typeString(args[0].Type())), nil
+		return cty.StringVal(TypeString(args[0].Type())), nil
 	},
 })